--- midix/src/parser/types/track_message/meta/mod.rs
@@ -1,6 +1,0 @@
-mod tempo;
-pub use tempo::*;
-mod tempo_sig;
-pub use tempo_sig::*;
-mod key_sig;
-pub use key_sig::*;
